+++ includes/database/educare-database.php
@@ -83,7 +83,7 @@ function educare_database_table($db = null) {
 	$table2 = "CREATE TABLE $Educare_results (
 		`id` mediumint(11) NOT NULL AUTO_INCREMENT,
 		`Name` varchar(80) NOT NULL,
-		`Roll_No` varchar(80) NOT NULL,
+		`Date_Of_Birth` varchar(80) NOT NULL,
 		`Regi_No` varchar(80) NOT NULL,
 		`Class` varchar(80) NOT NULL,
 		`Exam` varchar(80) NOT NULL,
@@ -102,7 +102,7 @@ function educare_database_table($db = null) {
   $table3 = "CREATE TABLE $Educare_students (
 		`id` mediumint(11) NOT NULL AUTO_INCREMENT,
 		`Name` varchar(80) NOT NULL,
-		`Roll_No` varchar(80) NOT NULL,
+		`Date_Of_Birth` varchar(80) NOT NULL,
 		`Regi_No` varchar(80) NOT NULL,
 		`Class` varchar(80) NOT NULL,
 		`Year` varchar(80) NOT NULL,